prov/efa: merge rxr_pkt_type_base.c/h into efa_rdm_pke_utils.c/h
Move functions in rxr_pkt_type_base.c to efa_rdm_pke_utils.c,
and rename

	rxr_pkt_connid_ptr()
	rxr_pkt_init_data_from_ope()

to
	efa_rdm_pke_connid_ptr()
	efa_rdm_pke_init_payload_from_ope()

also rename
	efa_rdm_pke_copy_payload()
to
	efa_rdm_pke_copy_payload_to_ope()
