+++ libfabric.vcxproj
@@ -886,7 +886,6 @@
     <ClCompile Include="prov\efa\src\dgram\efa_dgram_cq.c" />
     <ClCompile Include="prov\efa\src\dgram\efa_dgram_msg.c" />
     <ClCompile Include="prov\efa\src\dgram\efa_dgram_rma.c" />
-    <ClCompile Include="prov\efa\src\rdm\rxr_pkt_type_base.c" />
     <ClCompile Include="prov\efa\src\rdm\efa_rdm_ope.c" />
     <ClCompile Include="prov\efa\src\rdm\efa_rdm_atomic.c" />
     <ClCompile Include="prov\efa\src\rdm\efa_rdm_ep_utils.c" />
@@ -1014,7 +1013,7 @@
     <ClInclude Include="prov\efa\src\efa_base_ep.h" />
     <ClInclude Include="prov\efa\src\efa_cntr.h" />
     <ClInclude Include="prov\efa\src\dgram\efa_dgram.h" />
-    <ClInclude Include="prov\efa\src\rdm\rxr_pkt_type_base.h" />
+    <ClInclude Include="prov\efa\src\rdm\efa_rdm_pke_utils.h" />
     <ClInclude Include="prov\efa\src\rdm\rxr_pkt_type_misc.h" />
     <ClInclude Include="prov\efa\src\rdm\efa_rdm_atomic.h" />
     <ClInclude Include="prov\efa\src\rdm\efa_rdm_msg.h" />
