+++ prov/efa/src/rdm/efa_rdm_pke_utils.h
@@ -77,4 +77,15 @@ size_t efa_rdm_pke_get_segment_offset(struct efa_rdm_pke *pke)
 
 size_t efa_rdm_pke_get_payload_offset(struct efa_rdm_pke *pkt_entry);
 
+ssize_t efa_rdm_pke_init_payload_from_ope(struct efa_rdm_pke *pke,
+					  struct efa_rdm_ope *ope,
+					  size_t segment_offset,
+					  size_t payload_offset,
+					  size_t data_size);
+
+ssize_t efa_rdm_pke_copy_payload_to_ope(struct efa_rdm_pke *pke,
+					struct efa_rdm_ope *ope);
+
+uint32_t *efa_rdm_pke_connid_ptr(struct efa_rdm_pke *pkt_entry);
+
 #endif
