--- prov/efa/src/rdm/rxr_pkt_type_base.h
@@ -1,53 +1,0 @@
-/*
- * Copyright (c) 2021 Amazon.com, Inc. or its affiliates.
- * All rights reserved.
- *
- * This software is available to you under a choice of one of two
- * licenses.  You may choose to be licensed under the terms of the GNU
- * General Public License (GPL) Version 2, available from the file
- * COPYING in the main directory of this source tree, or the
- * BSD license below:
- *
- *     Redistribution and use in source and binary forms, with or
- *     without modification, are permitted provided that the following
- *     conditions are met:
- *
- *      - Redistributions of source code must retain the above
- *        copyright notice, this list of conditions and the following
- *        disclaimer.
- *
- *      - Redistributions in binary form must reproduce the above
- *        copyright notice, this list of conditions and the following
- *        disclaimer in the documentation and/or other materials
- *        provided with the distribution.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
- * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
- * BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
- * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-#ifndef _RXR_PKT_TYPE_BASE_H
-#define _RXR_PKT_TYPE_BASE_H
-
-
-
-uint32_t *rxr_pkt_connid_ptr(struct efa_rdm_pke *pkt_entry);
-
-int rxr_pkt_init_data_from_ope(struct efa_rdm_ep *ep,
-				    struct efa_rdm_pke *pkt_entry,
-				    size_t pkt_data_offset,
-				    struct efa_rdm_ope *ope,
-				    size_t op_data_offset,
-				    size_t data_size);
-
-ssize_t efa_rdm_pke_copy_payload(struct efa_rdm_pke *pke,
-				 struct efa_rdm_ope *ope);
-
-size_t rxr_pkt_data_size(struct efa_rdm_pke *pkt_entry);
-
-#endif
